backend/requirements.txt
flask
flask-cors
python-dotenv
replicate
